[examples] Inline input building and make test/benchmark/accuracy methods
Inline each variant's input construction into its own _make_inputs and drop
the make_*_inputs helpers from linear_attention_utils (make_mamba2_inputs
stays; mamba2 is off-harness). Make test/benchmark/accuracy methods on
LinearAttentionVariant with the shapes as fields.
