+++ examples/linear/example_retention.py
@@ -8,6 +8,7 @@ chunk_retention computes its own internal decay, so we do NOT pass g to FLA.
 
 from __future__ import annotations
 
+import math
 from typing import Any
 
 import torch
@@ -15,11 +16,7 @@ import torch
 from .linear_attention_engine import chunked_linear_attn
 from .linear_attention_harness import Inputs
 from .linear_attention_harness import LinearAttentionVariant
-from .linear_attention_harness import run_accuracy as _run_accuracy
-from .linear_attention_harness import run_benchmark as _run_benchmark
-from .linear_attention_harness import run_test as _run_test
 from .linear_attention_utils import chunked_linear_attn_reference
-from .linear_attention_utils import make_retention_inputs
 from .linear_attention_utils import naive_recurrent_reference
 from helion._testing import DEVICE
 
@@ -49,10 +46,16 @@ def _make_inputs(
     device: str | torch.device = DEVICE,
     requires_grad: bool = False,
 ) -> Inputs:
-    q, k, v, g, scale = make_retention_inputs(
-        B, H, T, D, DV, dtype=dtype, device=device, requires_grad=requires_grad
+    q = torch.randn(B, H, T, D, device=device, dtype=dtype, requires_grad=requires_grad)
+    k = torch.randn(B, H, T, D, device=device, dtype=dtype, requires_grad=requires_grad)
+    v = torch.randn(
+        B, H, T, DV, device=device, dtype=dtype, requires_grad=requires_grad
     )
-    return Inputs(q=q, k=k, v=v, scale=scale, g=g)
+    g_gamma = (
+        1 - 2.0 ** (-5 - torch.arange(H, dtype=torch.float32, device=device))
+    ).log()
+    g = g_gamma[None, :, None].expand(B, H, T).contiguous().to(dtype)
+    return Inputs(q=q, k=k, v=v, scale=1.0 / math.sqrt(D), g=g)
 
 
 def _helion_fwd(i: Inputs, C: int) -> torch.Tensor:
@@ -74,7 +77,7 @@ def _fla_fb(i: Inputs, go_t: torch.Tensor, scale: float) -> None:
 
 
 def _reference(i: Inputs) -> torch.Tensor:
-    # fp32 gamma: the per-head decay (~0.99997) needs fp32 to stay accurate
+    # fp32 gamma: the per-head decay (~0.97-0.9997) needs fp32 to stay accurate
     # over long T.
     assert i.g is not None
     return naive_recurrent_reference(i.q, i.k, i.v, i.g.float(), q_scale=i.scale)
@@ -93,33 +96,23 @@ VARIANT = LinearAttentionVariant(
     helion_fb=_helion_fb,
     reference=_reference,
     chunked_reference=_chunked_reference,
+    test_shape=(B, H, T, D, DV),
+    C=C,
+    bench_configs=BENCH_CONFIGS,
+    bench_C=BENCH_C,
     fla_fwd=_fla_fwd if _fla_chunk else None,
     fla_fb=_fla_fb if _fla_chunk else None,
     dtype=DTYPE,
 )
 
-
 # Module API consumed by run_linattn.py: test / benchmark / accuracy.
-def test() -> None:
-    _run_test(VARIANT, (B, H, T, D, DV), C)
-
-
-def benchmark(
-    configs: list | None = None,
-) -> list[tuple[str, float, float, float, float]]:
-    return _run_benchmark(
-        VARIANT, configs if configs is not None else BENCH_CONFIGS, BENCH_C
-    )
-
-
-def accuracy(configs: list | None = None) -> list[tuple[bool, bool]]:
-    return _run_accuracy(
-        VARIANT, configs if configs is not None else BENCH_CONFIGS, BENCH_C
-    )
+test = VARIANT.test
+benchmark = VARIANT.benchmark
+accuracy = VARIANT.accuracy
 
 
 def main() -> None:
-    print("=== Retention ===")
+    print(f"=== {VARIANT.title} ===")
     test()
     print()
     benchmark()
