+++ examples/linear/linear_attention_utils.py
@@ -12,8 +12,6 @@ Utility functions for the linear attention engine:
 
 from __future__ import annotations
 
-import math
-
 import torch
 import torch.nn.functional as F
 
@@ -357,165 +355,6 @@ def recurrent_step_reference(
 # ════════════════════════════════════════════════════════════════════════════════
 # Input generators for each attention variant
 # ════════════════════════════════════════════════════════════════════════════════
-
-
-def make_simple_gla_inputs(
-    B: int,
-    H: int,
-    T: int,
-    D: int,
-    DV: int,
-    dtype: torch.dtype = torch.bfloat16,
-    device: str | torch.device = "cuda",
-    requires_grad: bool = False,
-) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor, float]:
-    """Create inputs for simple GLA (scalar decay)."""
-    q = torch.randn(B, H, T, D, device=device, dtype=dtype, requires_grad=requires_grad)
-    k = torch.randn(B, H, T, D, device=device, dtype=dtype, requires_grad=requires_grad)
-    v = torch.randn(
-        B, H, T, DV, device=device, dtype=dtype, requires_grad=requires_grad
-    )
-    g = F.logsigmoid(torch.randn(B, H, T, device=device, dtype=dtype))
-    scale = 1.0 / math.sqrt(D)
-    return q, k, v, g, scale
-
-
-def make_full_gla_inputs(
-    B: int,
-    H: int,
-    T: int,
-    D: int,
-    DV: int,
-    dtype: torch.dtype = torch.bfloat16,
-    device: str | torch.device = "cuda",
-    requires_grad: bool = False,
-) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor, float]:
-    """Create inputs for full GLA (diagonal decay)."""
-    q = torch.randn(B, H, T, D, device=device, dtype=dtype, requires_grad=requires_grad)
-    k = torch.randn(B, H, T, D, device=device, dtype=dtype, requires_grad=requires_grad)
-    v = torch.randn(
-        B, H, T, DV, device=device, dtype=dtype, requires_grad=requires_grad
-    )
-    g = F.logsigmoid(torch.randn(B, H, T, D, device=device, dtype=dtype))
-    scale = 1.0 / math.sqrt(D)
-    return q, k, v, g, scale
-
-
-def make_delta_rule_inputs(
-    B: int,
-    H: int,
-    T: int,
-    D: int,
-    DV: int,
-    dtype: torch.dtype = torch.bfloat16,
-    device: str | torch.device = "cuda",
-    requires_grad: bool = False,
-) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor, float]:
-    """Create inputs for DeltaNet (correction, no decay)."""
-    q = torch.randn(B, H, T, D, device=device, dtype=dtype, requires_grad=requires_grad)
-    k = F.normalize(torch.randn(B, H, T, D, device=device, dtype=dtype), dim=-1)
-    if requires_grad:
-        k = k.detach().requires_grad_(True)
-    v = torch.randn(
-        B, H, T, DV, device=device, dtype=dtype, requires_grad=requires_grad
-    )
-    beta = torch.sigmoid(torch.randn(B, H, T, device=device, dtype=dtype))
-    g = torch.zeros(B, H, T, device=device, dtype=dtype)
-    scale = 1.0 / math.sqrt(D)
-    return q, k, v, g, beta, scale
-
-
-def make_gated_delta_rule_inputs(
-    B: int,
-    H: int,
-    T: int,
-    D: int,
-    DV: int,
-    dtype: torch.dtype = torch.bfloat16,
-    device: str | torch.device = "cuda",
-    requires_grad: bool = False,
-) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor, float]:
-    """Create inputs for Gated DeltaNet (correction + scalar decay)."""
-    q = torch.randn(B, H, T, D, device=device, dtype=dtype, requires_grad=requires_grad)
-    k = F.normalize(torch.randn(B, H, T, D, device=device, dtype=dtype), dim=-1)
-    if requires_grad:
-        k = k.detach().requires_grad_(True)
-    v = torch.randn(
-        B, H, T, DV, device=device, dtype=dtype, requires_grad=requires_grad
-    )
-    beta = torch.sigmoid(torch.randn(B, H, T, device=device, dtype=dtype))
-    g = F.logsigmoid(torch.randn(B, H, T, device=device, dtype=dtype))
-    scale = 1.0 / math.sqrt(D)
-    return q, k, v, g, beta, scale
-
-
-def make_kda_inputs(
-    B: int,
-    H: int,
-    T: int,
-    D: int,
-    DV: int,
-    dtype: torch.dtype = torch.bfloat16,
-    device: str | torch.device = "cuda",
-    requires_grad: bool = False,
-) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor, float]:
-    """Create inputs for KDA (diagonal decay + correction)."""
-    q = torch.randn(B, H, T, D, device=device, dtype=dtype, requires_grad=requires_grad)
-    k = F.normalize(torch.randn(B, H, T, D, device=device, dtype=dtype), dim=-1)
-    if requires_grad:
-        k = k.detach().requires_grad_(True)
-    v = torch.randn(
-        B, H, T, DV, device=device, dtype=dtype, requires_grad=requires_grad
-    )
-    g = -torch.rand(B, H, T, D, device=device, dtype=dtype).abs() * 0.1
-    beta = torch.sigmoid(torch.randn(B, H, T, device=device, dtype=dtype))
-    scale = 1.0 / math.sqrt(D)
-    return q, k, v, g, beta, scale
-
-
-def make_vanilla_linear_attn_inputs(
-    B: int,
-    H: int,
-    T: int,
-    D: int,
-    DV: int,
-    dtype: torch.dtype = torch.bfloat16,
-    device: str | torch.device = "cuda",
-    requires_grad: bool = False,
-) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor, float]:
-    """Create inputs for vanilla linear attention (no decay)."""
-    q = torch.randn(B, H, T, D, device=device, dtype=dtype, requires_grad=requires_grad)
-    k = torch.randn(B, H, T, D, device=device, dtype=dtype, requires_grad=requires_grad)
-    v = torch.randn(
-        B, H, T, DV, device=device, dtype=dtype, requires_grad=requires_grad
-    )
-    g = torch.zeros(B, H, T, device=device, dtype=dtype)
-    scale = 1.0 / math.sqrt(D)
-    return q, k, v, g, scale
-
-
-def make_retention_inputs(
-    B: int,
-    H: int,
-    T: int,
-    D: int,
-    DV: int,
-    dtype: torch.dtype = torch.bfloat16,
-    device: str | torch.device = "cuda",
-    requires_grad: bool = False,
-) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor, float]:
-    """Create inputs for RetNet (fixed per-head decay)."""
-    q = torch.randn(B, H, T, D, device=device, dtype=dtype, requires_grad=requires_grad)
-    k = torch.randn(B, H, T, D, device=device, dtype=dtype, requires_grad=requires_grad)
-    v = torch.randn(
-        B, H, T, DV, device=device, dtype=dtype, requires_grad=requires_grad
-    )
-    g_gamma = (
-        1 - 2.0 ** (-5 - torch.arange(H, dtype=torch.float32, device=device))
-    ).log()
-    g = g_gamma[None, :, None].expand(B, H, T).contiguous().to(dtype)
-    scale = 1.0 / math.sqrt(D)
-    return q, k, v, g, scale
 
 
 def make_mamba2_inputs(
